pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "illli-tts"
version = "0.1.0"
description = "基于 Qwen3-TTS 的 CLI 文字转语音工具"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "illli Studio", email = "studio@illli.com"}
]
keywords = ["tts", "qwen", "text-to-speech", "voice", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "torch>=2.0",
    "transformers>=4.30",
    "modelscope>=1.9.0",
    "soundfile>=0.12.0",
    "scipy>=1.11.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "httpx>=0.24",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
illli-tts = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
